add paramater to allow cache to be deleted or not when a refresh is done (you don't want to delete the cache on disk in the case of Symbian SDKs b/c if the IDE quits abnormally the cache in memory is not written to disk.
--HG--
branch : C3_BUILDER_WORK

+++ core/com.nokia.cpp.utils.core/src/com/nokia/cpp/internal/api/utils/core/CacheUtils.java
@@ -102,11 +102,13 @@ public class CacheUtils {
 		return result;		
 	}
 
-	public void removeCache(String identifier) {
+	public void removeCache(String identifier, boolean removeFromDisk) {
 		CacheEntry cache = caches.get(identifier);
 		if (cache != null) {
 			caches.remove(identifier);
-			cache.delete();
+			if (removeFromDisk){
+				cache.delete();
+			}
 		}
 	}
